switch to wasm32; no longer compiles (expected)
C code no longer compiles to a binary, because there is no libX11 for
wasm32 and we are no longer linking to the standard library.
Conveniently, clang wasm-ld dumps 553 lines about in which file what
undefined symbol is used.
This will be very helpful to remove some features from DOOM to get
the amount of missing symbols down!

But the `liblinuxxdoom.a` still compiles! We just need to provide the
missing functions during link time.

The rust code compiles into a wasm32 file.
But this file cannot yet be executed, since it requires a lot of
imports to be provided.

I modified `cargo run` to dump the imports we need.
This currently gives:

  (import "env" "XKeycodeToKeysym" (func $XKeycodeToKeysym (type 1)))
  (import "env" "XNextEvent" (func $XNextEvent (type 5)))
  (import "env" "XCreatePixmap" (func $XCreatePixmap (type 6)))
  (import "env" "XCreateGC" (func $XCreateGC (type 7)))
  (import "env" "XFillRectangle" (func $XFillRectangle (type 8)))
  (import "env" "XCreatePixmapCursor" (func $XCreatePixmapCursor (type 8)))
  (import "env" "XFreePixmap" (func $XFreePixmap (type 5)))
  (import "env" "XFreeGC" (func $XFreeGC (type 5)))
  (import "env" "XPending" (func $XPending (type 3)))
  (import "env" "XWarpPointer" (func $XWarpPointer (type 9)))
  (import "env" "malloc" (func $malloc (type 3)))
  (import "env" "gettimeofday" (func $gettimeofday (type 5)))
  (import "env" "exit" (func $exit (type 2)))
    [... truncated ... ]
  (export "memory" (memory 0))
  (export "main" (func $main))
  (export "__data_end" (global 1))
  (export "__heap_base" (global 2))

73 missing imports.

All the X... imports can probably be deleted when we write our own
graphics driver.

A few C standard library functions also need to be implemented.

I hope to get rid of all the networking stuff by dropping network
support for now.

Implementing the filesytem functions sounds complicated, but we only
need to support opening and reading doom1.wad, all other functions
can fail.

Maybe I can copy some functions such as `strncpy` from a small libc
implementation, such as musl or dietlibc.