add musl config.mak
config.mak was generated by configure tool. We should not Git ignore
this file.

Without config.mak, the make target lib/libc.a will fail.

Note that when we run the configure tool, it failed and thus config.mak
was missing. We actually copied this file from the original project
source code.