printing now works
Geting C printf & puts to work.
I did not want to change how the musl libc defaults on a `FILE *`
abstraction that's why I simply implement the `writev` syscall in
Rust, but only for `STDOUT` and `STDERR`, not for arbitrary files.

So, now Doom is starting untill the first malloc.