don't busy wait for a new game tick inside Doom, return control to browser sooner
Problem:
The game runs at ~35 FPS on my machine, but Chrome performance
debugging tools still show many dropped frames.

This is because the browser wants to animate at 60 FPS. In addition,
since Doom is still polling the time to know when it can proceed,
this is super energy inefficient and gives the browser no room for
background tasks, such as garbage collection.

Now, Doom still runs at ~35 FPS (this is what Doom was designed for),
but the browser gets a chance to render its 60 animation frames per
second and the system is mostly idle otherwise.

Optimizations:
Make Doom to return immediately when running one step of its game
loop if there is nothing to do, giving control back to the browser.

This improves the following:
* The animation frames per second went up from the ~35 to the target
  of 60.
* My CPU fan stops spinning at full speed when playing doom in the
  browser, since the browser is better at controlling animations than
  busy waiting for time to progress.
* The chrome developer console shows almost no more dropped frames.
* The gameplay feels smoother, since there are no longer any hiccups.
  Do we give the browser more room for GC?
* Firefox performance debugger now says: most "expensive function"
  with over 70% of time consumed is the browser's idle function! \m/